+++ .env.example
@@ -16,7 +16,7 @@ MAGIC_LOGIN_SECRET=secret
 
 # VAPID keys for WebPush Notifications (generate with: 'docker compose exec backend npx web-push generate-vapid-keys')
 VITE_PUBLIC_VAPID_KEY=''
-VITE_PRIVATE_VAPID_KEY=''
+PRIVATE_VAPID_KEY=''
 
 # Ports
 FRONTEND_PORT=5000
