+++ backend/app.ts
@@ -1,13 +1,11 @@
-import MongoStore from 'connect-mongo'
 import cors from 'cors'
 import express, { Request as ExRequest, Response as ExResponse } from 'express'
 import { rateLimit } from 'express-rate-limit'
 import session from 'express-session'
-import mongoose from 'mongoose'
 import swaggerUi from 'swagger-ui-express'
 import auth from './auth.js'
 import { errorHandler, RateLimitExceededError } from './controller/error.js'
-import { connectDB } from './db.js'
+import { connectDB, sessionStore } from './db.js'
 import { RegisterRoutes } from './dist/routes.js'
 import swaggerDocument from './dist/swagger.json' with { type: 'json' }
 import i18n from './i18n.js'
@@ -53,8 +51,6 @@ if (process.env.RATE_LIMIT_WINDOW_MS && process.env.RATE_LIMIT) {
   )
 }
 
-//makes sessionStore avaiable in other files aswell
-export const sessionStore = MongoStore.create({ client: mongoose.connection.getClient() })
 app.use(
   session({
     store: sessionStore,
