+++ backend/authStrategies/magiclogin.ts
@@ -2,8 +2,8 @@ import MagicLoginStrategy from 'passport-magic-login'
 import { escapeRegExp } from '../../common/scripts.js'
 import { NotAllowedError } from '../controller/error.js'
 import i18n from '../i18n.js'
-import { sendMail } from '../mail/mail.js'
 import User from '../models/user.js'
+import { sendMail } from '../notifications/mail.js'
 const secret = process.env.MAGIC_LOGIN_SECRET
 const callbackUrl = process.env.VITE_BACKEND_URL + '/auth/magiclogin/callback'
 const options = {
