+++ backend/pdf/helper.ts
@@ -16,9 +16,9 @@ import {
 } from '../../common/types.js'
 import { getConnectionSettings } from '../db.js'
 import i18n, { formatter } from '../i18n.js'
-import { getClient } from '../mail/mail.js'
 import DocumentFile from '../models/documentFile.js'
 import Organisation from '../models/organisation.js'
+import { getClient } from '../notifications/mail.js'
 import { generateAdvanceReport } from './advance.js'
 import { generateExpenseReportReport } from './expenseReport.js'
 import { generateHealthCareCostReport } from './healthCareCost.js'
