--- common/forms/user.json
@@ -1,1 +1,0 @@
-{"fk":{"rules":["required"],"label":{"de":"Authentifizierung","en":"Authentication"},"type":"object","schema":{"microsoft":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"placeholder":{"de":"Microsoft ID","en":"Microsoft ID"},"type":"text"},"ldapauth":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"placeholder":{"de":"LDAP UID","en":"LDAP UID"},"type":"text"},"magiclogin":{"rules":["nullable","email"],"columns":{"lg":{"container":4},"sm":{"container":6}},"placeholder":{"de":"Magic Login Email","en":"Magic Login Email"},"type":"text"}}},"email":{"rules":["required","email"],"placeholder":{"de":"Email Adresse","en":"Email address"},"type":"text"},"name":{"rules":["required"],"label":{"de":"Name","en":"Name"},"type":"object","schema":{"givenName":{"rules":["required"],"columns":{"lg":{"container":6},"sm":{"container":6}},"placeholder":{"de":"Vorname","en":"Given name"},"type":"text"},"familyName":{"rules":["required"],"columns":{"lg":{"container":6},"sm":{"container":6}},"placeholder":{"de":"Nachname","en":"Family name"},"type":"text"}}},"access":{"rules":["nullable"],"label":{"de":"Zugriffsrechte","en":"Access"},"type":"object","schema":{"user":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":true,"type":"checkbox","text":{"de":"Home","en":"Home"}},"inWork:expenseReport":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":true,"type":"checkbox","text":{"de":"Auslagenabrechnung initiieren","en":"Initiate Expenses"}},"inWork:healthCareCost":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":true,"type":"checkbox","text":{"de":"Krankenkostenabrechnung initiieren","en":"Initiate Health Care Costs"}},"appliedFor:travel":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":true,"type":"checkbox","text":{"de":"Reisekostenabrechnung beantragen","en":"Apply For Travels"}},"approved:travel":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":false,"type":"checkbox","text":{"de":"Reisekostenabrechnung ohne Genehmigung initiieren","en":"Initiate Travel Expenses Without Approval"}},"approve/travel":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":false,"type":"checkbox","text":{"de":"Reisen genehmigen","en":"Approve Travels"}},"examine/travel":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":false,"type":"checkbox","text":{"de":"Reisen prüfen","en":"Examine Travels"}},"examine/expenseReport":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":false,"type":"checkbox","text":{"de":"Auslagen prüfen","en":"Examine Expenses"}},"examine/healthCareCost":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":false,"type":"checkbox","text":{"de":"Krankenkosten prüfen","en":"Examine Health Care Costs"}},"confirm/healthCareCost":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":false,"type":"checkbox","text":{"de":"Krankenkosten bestätigen","en":"Confirm Health Care Costs"}},"admin":{"rules":["nullable"],"columns":{"lg":{"container":4},"sm":{"container":6}},"default":false,"type":"checkbox","text":{"de":"Administrator","en":"Administrator"}}}},"loseAccessAt":{"rules":["nullable"],"info":{"de":"Stelle hier ein Datum ein, ab dem der Nutzer keinen Zugriff mehr zu abrechnung🧾 haben soll.","en":"Set a date here from which the user should no longer have access to abrechnung🧾"},"type":"date","time":false,"placeholder":{"de":"Ablaufdatum","en":"Expiration date"}},"settings":{"rules":["required"],"label":{"de":"Einstellungen","en":"Settings"},"type":"object","schema":{"language":{"rules":["nullable"],"default":"de","placeholder":{"de":"Sprache","en":"Language"},"type":"select","items":{"de":{"de":"Deutsch 🇩🇪","en":"German 🇩🇪"},"en":{"de":"Englisch 🇬🇧","en":"English 🇬🇧"}}},"lastCurrencies":{"rules":["min:0"],"type":"currency","extendOptions":{"mode":"multiple"},"placeholder":{"de":"Letzte Währungen","en":"Last currencies"}},"lastCountries":{"rules":["min:0"],"type":"country","extendOptions":{"mode":"multiple"},"placeholder":{"de":"Letzte Länder","en":"Last countries"}},"projects":{"rules":["min:0"],"type":"project","extendOptions":{"mode":"multiple"},"placeholder":{"de":"Projekte","en":"Projects"}},"insurance":{"rules":["nullable"],"type":"healthinsurance","placeholder":{"de":"Versicherung","en":"Health insurance"}},"organisation":{"rules":["nullable"],"type":"organisation","placeholder":{"de":"Organisation","en":"Organisation"}},"showInstallBanner":{"rules":["nullable"],"default":true,"type":"installbanner","placeholder":{"de":"labels.showInstallBanner","en":"labels.showInstallBanner"}}}},"token":{"rules":["nullable"],"type":"token","placeholder":{"de":"labels.token","en":"labels.token"}}}
