+++ frontend/src/components/HomePage.vue
@@ -44,7 +44,7 @@
           <h1>{{ $t('headlines.home') }}</h1>
         </div>
         <div v-if="!$root.settings.disableReportType.travel && $root.user.access['appliedFor:travel']" class="col-auto">
-          <button class="btn btn-secondary" @click="showModal('add', {}, 'travel')" :disabled="$root.isOffline">
+          <button class="btn btn-secondary" @click="showModal('add', {}, 'travel')">
             <i class="bi bi-plus-lg"></i>
             <span class="ms-1">{{
               $t($root.user.access['approved:travel'] ? 'labels.addX' : 'labels.applyForX', { X: $t('labels.travel') })
@@ -52,13 +52,13 @@
           </button>
         </div>
         <div v-if="!$root.settings.disableReportType.expenseReport && $root.user.access['inWork:expenseReport']" class="col-auto">
-          <button class="btn btn-secondary" @click="showModal('add', {}, 'expenseReport')" :disabled="$root.isOffline">
+          <button class="btn btn-secondary" @click="showModal('add', {}, 'expenseReport')">
             <i class="bi bi-plus-lg"></i>
             <span class="ms-1">{{ $t('labels.addX', { X: $t('labels.expenseReport') }) }}</span>
           </button>
         </div>
         <div v-if="!$root.settings.disableReportType.healthCareCost && $root.user.access['inWork:healthCareCost']" class="col-auto">
-          <button class="btn btn-secondary" @click="showModal('add', {}, 'healthCareCost')" :disabled="$root.isOffline">
+          <button class="btn btn-secondary" @click="showModal('add', {}, 'healthCareCost')">
             <i class="bi bi-plus-lg"></i>
             <span class="ms-1">{{ $t('labels.submitX', { X: $t('labels.healthCareCost') }) }}</span>
           </button>
